add new file for temp friend view
